--- backend/alembic/versions/022d04479a33_added_roles_to_social_group_members.py
@@ -1,34 +1,0 @@
-"""added roles to social group members
-
-Revision ID: 022d04479a33
-Revises: 4a508ba69d73
-Create Date: 2020-07-19 20:04:31.235947
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = '022d04479a33'
-down_revision = '4a508ba69d73'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('social_group_members', sa.Column('role', sa.Enum('ADMIN', 'MEMBER', name='socialgrouprole'), nullable=False))
-    op.drop_constraint('social_groups_ibfk_1', 'social_groups', type_='foreignkey')
-    op.drop_column('social_groups', 'owner_id')
-    op.drop_column('social_groups', 'admins')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('social_groups', sa.Column('admins', mysql.JSON(), nullable=True))
-    op.add_column('social_groups', sa.Column('owner_id', mysql.BIGINT(), autoincrement=False, nullable=False))
-    op.create_foreign_key('social_groups_ibfk_1', 'social_groups', 'users', ['owner_id'], ['id'])
-    op.drop_column('social_group_members', 'role')
-    # ### end Alembic commands ###
