--- backend/alembic/versions/0c379ced6391_added_names_to_tags.py
@@ -1,28 +1,0 @@
-"""added names to tags
-
-Revision ID: 0c379ced6391
-Revises: 022d04479a33
-Create Date: 2020-07-19 20:37:53.926754
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '0c379ced6391'
-down_revision = '022d04479a33'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('tags', sa.Column('name', sa.String(length=255), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('tags', 'name')
-    # ### end Alembic commands ###
