--- backend/alembic/versions/3ea635ed847d_names_for_user_and_groups.py
@@ -1,34 +1,0 @@
-"""names for user and groups
-
-Revision ID: 3ea635ed847d
-Revises: 9cfd23b0bc71
-Create Date: 2020-07-19 22:29:51.012591
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '3ea635ed847d'
-down_revision = '9cfd23b0bc71'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('social_groups', sa.Column('name', sa.String(length=255), nullable=False))
-    op.create_unique_constraint('uix_name', 'social_groups', ['name'])
-    op.add_column('users', sa.Column('name', sa.String(length=255), nullable=False))
-    op.create_unique_constraint('uix_name', 'users', ['name'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('uix_name', 'users', type_='unique')
-    op.drop_column('users', 'name')
-    op.drop_constraint('uix_name', 'social_groups', type_='unique')
-    op.drop_column('social_groups', 'name')
-    # ### end Alembic commands ###
