--- backend/alembic/versions/4a508ba69d73_init.py
@@ -1,101 +1,0 @@
-"""init
-
-Revision ID: 4a508ba69d73
-Revises: 
-Create Date: 2020-07-19 19:57:26.884807
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = '4a508ba69d73'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('uid', sa.String(length=255), nullable=False),
-    sa.Column('metadata_json', mysql.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('social_groups',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('metadata_json', mysql.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('owner_id', sa.BigInteger(), nullable=False),
-    sa.Column('admins', mysql.JSON(), nullable=True),
-    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('posts',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('metadata_json', mysql.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('social_group_id', sa.BigInteger(), nullable=False),
-    sa.Column('owner_id', sa.BigInteger(), nullable=False),
-    sa.Column('visibility', sa.Enum('PUBLIC', 'PRIVATE', name='visibilitytype'), nullable=False),
-    sa.Column('depth', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
-    sa.ForeignKeyConstraint(['social_group_id'], ['social_groups.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('social_group_members',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('user_id', sa.BigInteger(), nullable=False),
-    sa.Column('social_group_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['social_group_id'], ['social_groups.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('post_children',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('parent_post_id', sa.BigInteger(), nullable=False),
-    sa.Column('child_post_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['child_post_id'], ['posts.id'], ),
-    sa.ForeignKeyConstraint(['parent_post_id'], ['posts.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('tags',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('post_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['post_id'], ['posts.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('votes',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('vote_type', sa.Enum('DOWN', 'UP', name='votetype'), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('owner_id', sa.BigInteger(), nullable=False),
-    sa.Column('post_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
-    sa.ForeignKeyConstraint(['post_id'], ['posts.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('votes')
-    op.drop_table('tags')
-    op.drop_table('post_children')
-    op.drop_table('social_group_members')
-    op.drop_table('posts')
-    op.drop_table('social_groups')
-    op.drop_table('users')
-    # ### end Alembic commands ###
