+++ backend/alembic/versions/6fcb42694cbf_init.py
@@ -1,0 +1,105 @@
+"""init
+
+Revision ID: 6fcb42694cbf
+Revises: 
+Create Date: 2020-07-21 18:48:45.537997
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import mysql
+import common
+
+# revision identifiers, used by Alembic.
+revision = '6fcb42694cbf'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('social_groups',
+    sa.Column('id', common.models.GUID(length=32), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('metadata_json', mysql.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name', name='uix_name')
+    )
+    op.create_table('users',
+    sa.Column('id', common.models.GUID(length=32), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('uid', sa.String(length=255), nullable=False),
+    sa.Column('metadata_json', mysql.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name', name='uix_name')
+    )
+    op.create_table('posts',
+    sa.Column('id', common.models.GUID(length=32), nullable=False),
+    sa.Column('metadata_json', mysql.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.Column('social_group_id', common.models.GUID(length=32), nullable=False),
+    sa.Column('owner_id', common.models.GUID(length=32), nullable=False),
+    sa.Column('visibility', sa.Enum('PUBLIC', 'PRIVATE', name='visibilitytype'), nullable=False),
+    sa.Column('depth', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['social_group_id'], ['social_groups.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('social_group_members',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.Column('user_id', common.models.GUID(length=32), nullable=False),
+    sa.Column('social_group_id', common.models.GUID(length=32), nullable=False),
+    sa.Column('role', sa.Enum('ADMIN', 'MEMBER', name='socialgrouprole'), nullable=False),
+    sa.ForeignKeyConstraint(['social_group_id'], ['social_groups.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('post_children',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('parent_post_id', common.models.GUID(length=32), nullable=False),
+    sa.Column('child_post_id', common.models.GUID(length=32), nullable=False),
+    sa.ForeignKeyConstraint(['child_post_id'], ['posts.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['parent_post_id'], ['posts.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('tags',
+    sa.Column('id', common.models.GUID(length=32), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.Column('post_id', common.models.GUID(length=32), nullable=False),
+    sa.ForeignKeyConstraint(['post_id'], ['posts.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('votes',
+    sa.Column('id', common.models.GUID(length=32), nullable=False),
+    sa.Column('vote_type', sa.Enum('DOWN', 'UP', name='votetype'), nullable=False),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.Column('owner_id', common.models.GUID(length=32), nullable=False),
+    sa.Column('post_id', common.models.GUID(length=32), nullable=False),
+    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['post_id'], ['posts.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('votes')
+    op.drop_table('tags')
+    op.drop_table('post_children')
+    op.drop_table('social_group_members')
+    op.drop_table('posts')
+    op.drop_table('users')
+    op.drop_table('social_groups')
+    # ### end Alembic commands ###
