--- backend/alembic/versions/9cfd23b0bc71_delete_on_cascade.py
@@ -1,62 +1,0 @@
-"""delete on cascade
-
-Revision ID: 9cfd23b0bc71
-Revises: 0c379ced6391
-Create Date: 2020-07-19 20:49:20.026921
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '9cfd23b0bc71'
-down_revision = '0c379ced6391'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('post_children_ibfk_1', 'post_children', type_='foreignkey')
-    op.drop_constraint('post_children_ibfk_2', 'post_children', type_='foreignkey')
-    op.create_foreign_key(None, 'post_children', 'posts', ['parent_post_id'], ['id'], ondelete='CASCADE')
-    op.create_foreign_key(None, 'post_children', 'posts', ['child_post_id'], ['id'], ondelete='CASCADE')
-    op.drop_constraint('posts_ibfk_1', 'posts', type_='foreignkey')
-    op.drop_constraint('posts_ibfk_2', 'posts', type_='foreignkey')
-    op.create_foreign_key(None, 'posts', 'social_groups', ['social_group_id'], ['id'], ondelete='CASCADE')
-    op.create_foreign_key(None, 'posts', 'users', ['owner_id'], ['id'], ondelete='CASCADE')
-    op.drop_constraint('social_group_members_ibfk_2', 'social_group_members', type_='foreignkey')
-    op.drop_constraint('social_group_members_ibfk_1', 'social_group_members', type_='foreignkey')
-    op.create_foreign_key(None, 'social_group_members', 'users', ['user_id'], ['id'], ondelete='CASCADE')
-    op.create_foreign_key(None, 'social_group_members', 'social_groups', ['social_group_id'], ['id'], ondelete='CASCADE')
-    op.drop_constraint('tags_ibfk_1', 'tags', type_='foreignkey')
-    op.create_foreign_key(None, 'tags', 'posts', ['post_id'], ['id'], ondelete='CASCADE')
-    op.drop_constraint('votes_ibfk_1', 'votes', type_='foreignkey')
-    op.drop_constraint('votes_ibfk_2', 'votes', type_='foreignkey')
-    op.create_foreign_key(None, 'votes', 'users', ['owner_id'], ['id'], ondelete='CASCADE')
-    op.create_foreign_key(None, 'votes', 'posts', ['post_id'], ['id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'votes', type_='foreignkey')
-    op.drop_constraint(None, 'votes', type_='foreignkey')
-    op.create_foreign_key('votes_ibfk_2', 'votes', 'posts', ['post_id'], ['id'])
-    op.create_foreign_key('votes_ibfk_1', 'votes', 'users', ['owner_id'], ['id'])
-    op.drop_constraint(None, 'tags', type_='foreignkey')
-    op.create_foreign_key('tags_ibfk_1', 'tags', 'posts', ['post_id'], ['id'])
-    op.drop_constraint(None, 'social_group_members', type_='foreignkey')
-    op.drop_constraint(None, 'social_group_members', type_='foreignkey')
-    op.create_foreign_key('social_group_members_ibfk_1', 'social_group_members', 'social_groups', ['social_group_id'], ['id'])
-    op.create_foreign_key('social_group_members_ibfk_2', 'social_group_members', 'users', ['user_id'], ['id'])
-    op.drop_constraint(None, 'posts', type_='foreignkey')
-    op.drop_constraint(None, 'posts', type_='foreignkey')
-    op.create_foreign_key('posts_ibfk_2', 'posts', 'social_groups', ['social_group_id'], ['id'])
-    op.create_foreign_key('posts_ibfk_1', 'posts', 'users', ['owner_id'], ['id'])
-    op.drop_constraint(None, 'post_children', type_='foreignkey')
-    op.drop_constraint(None, 'post_children', type_='foreignkey')
-    op.create_foreign_key('post_children_ibfk_2', 'post_children', 'posts', ['parent_post_id'], ['id'])
-    op.create_foreign_key('post_children_ibfk_1', 'post_children', 'posts', ['child_post_id'], ['id'])
-    # ### end Alembic commands ###
